// frontend/src/types/api/listCouponsApiCouponsGetSort.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

export type ListCouponsApiCouponsGetSort = typeof ListCouponsApiCouponsGetSort[keyof typeof ListCouponsApiCouponsGetSort];


export const ListCouponsApiCouponsGetSort = {
  newest: 'newest',
  oldest: 'oldest',
  store: 'store',
  expiry: 'expiry',
} as const;